refactor: reorganiza estrutura do projeto
