--- src/main/java/Anao.java
@@ -1,5 +1,0 @@
-public interface Anao {
-  public int beber();
-
-  public String falar();
-}
