--- src/main/java/Elfo.java
@@ -1,5 +1,0 @@
-public interface Elfo {
-  public String falar();
-
-  public String falarElfico();
-}
