--- src/main/java/Humano.java
@@ -1,5 +1,0 @@
-public interface Humano {
-  public String falar();
-
-  public void envelhecer();
-}
