--- src/main/java/Maia.java
@@ -1,5 +1,0 @@
-public interface Maia {
-  public void ressucitar();
-
-  public String falar();
-}
