--- src/main/java/Monstro.java
@@ -1,3 +1,0 @@
-public interface Monstro {
-  public String grunhido();
-}
