--- src/main/java/PersonagemJaEstaNoMapaException.java
@@ -1,5 +1,0 @@
-public class PersonagemJaEstaNoMapaException extends Exception {
-  public PersonagemJaEstaNoMapaException(String mensagem) {
-    super(mensagem);
-  }
-}
