--- src/main/java/PersonagemNaoEncontradoNoMapaException.java
@@ -1,5 +1,0 @@
-public class PersonagemNaoEncontradoNoMapaException extends Exception {
-  public PersonagemNaoEncontradoNoMapaException(String mensagem) {
-    super(mensagem);
-  }
-}
