--- src/main/java/PosicaoOcupadaException.java
@@ -1,5 +1,0 @@
-public class PosicaoOcupadaException extends Exception {
-  public PosicaoOcupadaException(String mensagem) {
-    super(mensagem);
-  }
-}
