--- src/main/java/Saruman.java
@@ -1,21 +1,0 @@
-public class Saruman extends Mago implements Maia {
-  public Saruman() {
-    super("Saruman", 2, 2, 9, 70, false);
-  }
-
-  @Override
-  public String falar() {
-    return "Against the power of Mordor there can be no victory.";
-  }
-
-  @Override
-  public void ressucitar() {
-    if (this.getConstituicao() == 0) {
-      this.setConstituicao(this.getConstituicao());
-    }
-  }
-
-  public String toString() {
-    return "S";
-  }
-}
