--- src/main/java/SauronDominaOMundoException.java
@@ -1,5 +1,0 @@
-public class SauronDominaOMundoException extends Exception {
-  public SauronDominaOMundoException(String message) {
-    super(message);
-  }
-}
