+++ src/main/java/classes/Saruman.java
@@ -1,0 +1,26 @@
+package classes;
+
+import interfaces.Maia;
+import racas.Mago;
+
+public class Saruman extends Mago implements Maia {
+  public Saruman() {
+    super("Saruman", 2, 2, 9, 70, false);
+  }
+
+  @Override
+  public String falar() {
+    return "Against the power of Mordor there can be no victory.";
+  }
+
+  @Override
+  public void ressucitar() {
+    if (this.getConstituicao() == 0) {
+      this.setConstituicao(this.getConstituicao());
+    }
+  }
+
+  public String toString() {
+    return "S";
+  }
+}
