+++ src/main/java/exceptions/PersonagemJaEstaNoMapaException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class PersonagemJaEstaNoMapaException extends Exception {
+  public PersonagemJaEstaNoMapaException(String mensagem) {
+    super(mensagem);
+  }
+}
