+++ src/main/java/exceptions/PersonagemNaoEncontradoNoMapaException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class PersonagemNaoEncontradoNoMapaException extends Exception {
+  public PersonagemNaoEncontradoNoMapaException(String mensagem) {
+    super(mensagem);
+  }
+}
