+++ src/main/java/exceptions/PosicaoOcupadaException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class PosicaoOcupadaException extends Exception {
+  public PosicaoOcupadaException(String mensagem) {
+    super(mensagem);
+  }
+}
