+++ src/main/java/exceptions/SauronDominaOMundoException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class SauronDominaOMundoException extends Exception {
+  public SauronDominaOMundoException(String message) {
+    super(message);
+  }
+}
