+++ src/main/java/interfaces/Anao.java
@@ -1,0 +1,7 @@
+package interfaces;
+
+public interface Anao {
+  public int beber();
+
+  public String falar();
+}
