+++ src/main/java/interfaces/Elfo.java
@@ -1,0 +1,7 @@
+package interfaces;
+
+public interface Elfo {
+  public String falar();
+
+  public String falarElfico();
+}
