+++ src/main/java/interfaces/Humano.java
@@ -1,0 +1,7 @@
+package interfaces;
+
+public interface Humano {
+  public String falar();
+
+  public void envelhecer();
+}
