+++ src/main/java/interfaces/Maia.java
@@ -1,0 +1,7 @@
+package interfaces;
+
+public interface Maia {
+  public void ressucitar();
+
+  public String falar();
+}
