+++ src/main/java/interfaces/Monstro.java
@@ -1,0 +1,5 @@
+package interfaces;
+
+public interface Monstro {
+  public String grunhido();
+}
